pyproject.toml
[project]
name = "mcp-gmail-server"
version = "0.1.0"
description = "MCP server for Gmail API integration"
readme = "README.md"
license = "MIT"
authors = [
    { name = "cutsome", email = "nissikpu@gmail.com" }
]
keywords = ["mcp", "gmail", "api", "model-context-protocol"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
]
requires-python = ">=3.13"
dependencies = [
    "google-api-python-client>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "mcp>=0.1.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
mcp-gmail-server = "mcp_gmail_server.main:run"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[project.urls]
Homepage = "https://github.com/cutsome/gmail-mcp"
Repository = "https://github.com/cutsome/gmail-mcp"
Issues = "https://github.com/cutsome/gmail-mcp/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
